Refactor codebase for modularity and simplify DupeFinder API
- Split src/lib.rs into types, hashing, traversal, and finder modules.
- Consolidate async dependencies by moving to tokio::fs and tokio::io.
- Replace tokio-task-tracker and FuturesUnordered with tokio::task::JoinSet.
- Optimize hashing for small files (<= 4096 bytes) by skipping intermediate checks.
- Add find_duplicates method to DupeFinder returning an mpsc::Receiver for easier usage.
- Clean up main.rs and improve error handling.
